Start adding the app
With this commit i am starting to add the new Ikarus application. The
entire app has been rewritten completely from scratch. The old
"QR Code Scanner App" is gone.

I used this rewrite to modernize the codebase and switch some parts to
new "best practices". This includes using Jetpack for example.

This commit contains the gradle files for the app.

Signed-off-by: Leo Unglaub <leo@unglaub.at>